server/requirements.txt
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
